Redesign with Visitor pattern.
Each pojo element now contains a method which will use a ParserVisitor implementation to parse any kind of data into an object object of the same class. The ParserVisitor implementation will handle the conversion.
A Csv implementation will be made soon.

I plan to have a Json, java serialisation, and maybe a Sql implementation in the future. I'm mostly interested in the Csv and Json implementation as the file used as a database could potentially be stored on a remote server, and with required implementation, such database could be modified and accessed by a mobile application. I'm mostly looking at Android smartphones.

Right now the main idea is to have a budgeting application, which I can access on the terminal and remotely from a smartphone when I am outside.